+++ examples/typst-packages-unpublished/flake.nix
@@ -1,5 +1,5 @@
 {
-  description = "A Typst project that uses Typst packages";
+  description = "A Typst project that uses unpublished Typst packages";
 
   inputs = {
     nixpkgs.url = "github:NixOS/nixpkgs/nixos-unstable";
@@ -10,6 +10,12 @@
     };
 
     flake-utils.url = "github:numtide/flake-utils";
+
+    # TODO: Change this and the list in `unpublishedTypstPackages`
+    my-typst-package = {
+      url = "github:loqusion/my-typst-package";
+      flake = false;
+    };
 
     # Example of downloading icons from a non-flake source
     # font-awesome = {
@@ -27,6 +33,7 @@
     flake-utils.lib.eachDefaultSystem (system: let
       pkgs = nixpkgs.legacyPackages.${system};
       inherit (pkgs) lib;
+      inherit (lib.strings) escapeShellArg toShellVars;
 
       typixLib = typix.lib.${system};
 
@@ -48,13 +55,52 @@
         ];
       };
 
+      mkTypstPackageDrv = {
+        name,
+        version,
+        namespace,
+        input,
+        subdir ? "",
+      }: let
+        outSubdir = "${namespace}/${name}/${version}";
+      in
+        pkgs.stdenvNoCC.mkDerivation {
+          inherit name;
+          src = input;
+          dontBuild = true;
+          installPhase = ''
+            ${toShellVars {inherit subdir outSubdir;}}
+            mkdir -p $out/$outSubdir
+            if [ -n "$subdir" ]; then
+              if [ ! -e "$src/$subdir" ]; then
+                echo "error: subdir '$subdir' does not exist in $src" >&2
+                echo "contents of $src:" >&2
+                ls "$src" >&2
+                exit 1
+              fi
+              cp -r "$src/$subdir"/* -t $out/$outSubdir
+            else
+              cp -r $src/* -t $out/$outSubdir
+            fi
+          '';
+        };
+
+      unpublishedTypstPackages = pkgs.symlinkJoin {
+        name = "unpublished-typst-packages";
+        paths = map mkTypstPackageDrv [
+          # TODO: Change this
+          {
+            name = "my-typst-package";
+            version = "0.1.0";
+            namespace = "local";
+            input = inputs.my-typst-package;
+          }
+        ];
+      };
+
+      # Any transitive dependencies must be added here
+      # See https://loqusion.github.io/typix/recipes/using-typst-packages.html#the-typstpackages-attribute
       unstableTypstPackages = [
-        {
-          name = "cetz";
-          version = "0.3.4";
-          hash = "sha256-5w3UYRUSdi4hCvAjrp9HslzrUw7BhgDdeCiDRHGvqd4=";
-        }
-        # Required by cetz
         {
           name = "oxifmt";
           version = "0.2.1";
@@ -66,18 +112,25 @@
       # to the current directory
       build-drv = typixLib.buildTypstProject (commonArgs
         // {
-          inherit src unstableTypstPackages;
+          inherit src;
+          inherit unstableTypstPackages;
+          TYPST_PACKAGE_PATH = unpublishedTypstPackages;
         });
 
       # Compile a Typst project, and then copy the result
       # to the current directory
       build-script = typixLib.buildTypstProjectLocal (commonArgs
         // {
-          inherit src unstableTypstPackages;
+          inherit src;
+          inherit unstableTypstPackages;
+          TYPST_PACKAGE_PATH = unpublishedTypstPackages;
         });
 
       # Watch a project and recompile on changes
-      watch-script = typixLib.watchTypstProject commonArgs;
+      watch-script = typixLib.watchTypstProject (commonArgs
+        // {
+          typstWatchCommand = "TYPST_PACKAGE_PATH=${escapeShellArg unpublishedTypstPackages} typst watch";
+        });
     in {
       checks = {
         inherit build-drv build-script watch-script;